Split common WPF to separate project
