+++ OpenKh.Tools.Common.Wpf/MessageDialog.xaml
@@ -1,4 +1,4 @@
-<Window x:Class="OpenKh.Tools.Common.MessageDialog"
+<Window x:Class="OpenKh.Tools.Common.Wpf.MessageDialog"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
