+++ OpenKh.Tools.Common.Wpf/OpenKh.Tools.Common.Wpf.csproj
@@ -1,22 +1,25 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>WinExe</OutputType>
     <TargetFramework>net5.0-windows</TargetFramework>
     <LangVersion>latest</LangVersion>
     <UseWPF>true</UseWPF>
-    <AssemblyTitle>Image viewer</AssemblyTitle>
-    <Product>Image viewer - OpenKH</Product>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\OpenKh.Bbs\OpenKh.Bbs.csproj" />
-    <ProjectReference Include="..\OpenKh.Common\OpenKh.Common.csproj" />
-    <ProjectReference Include="..\OpenKh.Tools.Common.WPF\OpenKh.Tools.Common.Wpf.csproj" />
     <ProjectReference Include="..\OpenKh.Tools.Common\OpenKh.Tools.Common.csproj" />
+    <ProjectReference Include="..\XeEngine.Tools.Public\Xe.Drawing\Xe.Drawing.csproj" />
     <ProjectReference Include="..\XeEngine.Tools.Public\Xe.Tools.Wpf\Xe.Tools.Wpf.csproj" />
     <ProjectReference Include="..\XeEngine.Tools.Public\Xe.Tools\Xe.Tools.csproj" />
-    <ProjectReference Include="..\OpenKh.Command.ImgTool\OpenKh.Command.ImgTool.csproj" />
   </ItemGroup>
-  
+
+  <ItemGroup>
+    <Page Update="Controls\Timeline.xaml">
+      <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
+    </Page>
+    <Page Update="MessageDialog.xaml">
+      <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
+    </Page>
+  </ItemGroup>
+
 </Project>
