+++ OpenKh.Tools.Common.Wpf/Utilities.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Reflection;
 using System.Windows;
 
-namespace OpenKh.Tools.Common
+namespace OpenKh.Tools.Common.Wpf
 {
     public static class Utilities
     {
