+++ OpenKh.Tools.CtdEditor/Views/CtdView.xaml
@@ -1,10 +1,10 @@
-﻿<UserControl x:Class="OpenKh.Tools.CtdEditor.Views.CtdView"
+<UserControl x:Class="OpenKh.Tools.CtdEditor.Views.CtdView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:OpenKh.Tools.CtdEditor.Views"
-             xmlns:controls="clr-namespace:OpenKh.Tools.Common.Controls;assembly=OpenKh.Tools.Common"
+             xmlns:controls="clr-namespace:OpenKh.Tools.Common.Wpf.Controls;assembly=OpenKh.Tools.Common.Wpf"
              mc:Ignorable="d" 
              d:DesignHeight="350" d:DesignWidth="600">
     <Grid>
