+++ OpenKh.Tools.CtdEditor/Views/FontWindow.xaml
@@ -1,11 +1,11 @@
-﻿<Window x:Class="OpenKh.Tools.CtdEditor.Views.FontWindow"
+<Window x:Class="OpenKh.Tools.CtdEditor.Views.FontWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:OpenKh.Tools.CtdEditor.Views"
         xmlns:controls="clr-namespace:Xe.Tools.Wpf.Controls;assembly=Xe.Tools.Wpf"
-        xmlns:controls1="clr-namespace:OpenKh.Tools.Common.Controls;assembly=OpenKh.Tools.Common"
+        xmlns:controls1="clr-namespace:OpenKh.Tools.Common.Wpf.Controls;assembly=OpenKh.Tools.Common.Wpf"
         mc:Ignorable="d"
         Title="Font window" Height="250" Width="400" ResizeMode="NoResize">
     <Window.Resources>
