+++ OpenKh.Tools.Kh2BattleEditor/ViewModels/EnmpViewModel.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Windows;
 using OpenKh.Kh2;
 using OpenKh.Kh2.Battle;
-using OpenKh.Tools.Common.Models;
+using OpenKh.Tools.Common.Wpf.Models;
 using OpenKh.Tools.Kh2BattleEditor.Extensions;
 using OpenKh.Tools.Kh2BattleEditor.Interfaces;
 using OpenKh.Tools.Kh2BattleEditor.Services;
