+++ OpenKh.Tools.Kh2MapStudio/OpenKh.Tools.Kh2MapStudio.csproj
@@ -35,6 +35,7 @@
 		<ProjectReference Include="..\OpenKh.Engine\OpenKh.Engine.csproj" />
 		<ProjectReference Include="..\OpenKh.Kh2\OpenKh.Kh2.csproj" />
 		<ProjectReference Include="..\OpenKh.Tools.Common.CustomImGui\OpenKh.Tools.Common.CustomImGui.csproj" />
+		<ProjectReference Include="..\OpenKh.Tools.Common.WPF\OpenKh.Tools.Common.Wpf.csproj" />
 		<ProjectReference Include="..\OpenKh.Tools.Common\OpenKh.Tools.Common.csproj" />
 	</ItemGroup>
