+++ OpenKh.Tools.Kh2SystemEditor/ViewModels/FtstViewModel.cs
@@ -7,8 +7,8 @@ using System.Windows.Input;
 using System.Windows.Media;
 using OpenKh.Kh2;
 using OpenKh.Kh2.SystemData;
-using OpenKh.Tools.Common;
-using OpenKh.Tools.Common.Models;
+using OpenKh.Tools.Common.Wpf;
+using OpenKh.Tools.Common.Wpf.Models;
 using OpenKh.Tools.Kh2SystemEditor.Extensions;
 using OpenKh.Tools.Kh2SystemEditor.Interfaces;
 using Xe.Tools;
