+++ OpenKh.Tools.Kh2SystemEditor/ViewModels/MemtViewModel.cs
@@ -1,6 +1,7 @@
 using OpenKh.Kh2;
 using OpenKh.Kh2.SystemData;
 using OpenKh.Tools.Common.Models;
+using OpenKh.Tools.Common.Wpf.Models;
 using OpenKh.Tools.Kh2SystemEditor.Extensions;
 using OpenKh.Tools.Kh2SystemEditor.Interfaces;
 using OpenKh.Tools.Kh2SystemEditor.Models;
