+++ OpenKh.Tools.Kh2SystemEditor/ViewModels/TrsrViewModel.cs
@@ -5,6 +5,7 @@ using System.Windows;
 using OpenKh.Kh2;
 using OpenKh.Kh2.SystemData;
 using OpenKh.Tools.Common.Models;
+using OpenKh.Tools.Common.Wpf.Models;
 using OpenKh.Tools.Kh2SystemEditor.Extensions;
 using OpenKh.Tools.Kh2SystemEditor.Interfaces;
 using Xe.Tools;
